fix(canon): preserve relative paths in option probes
